Fix typo trafic->traffic (incl. renaming files/directories)
This is an API breaking commit.
