+++ doc/sphinxDoc/source/preprocessor/python_modules/actions.railway_traffic.rst
@@ -1,11 +1,12 @@
-Roadway traffic
+Railway traffic
 ===============
+
 
 
 IAPF rail load models
 ---------------------
 
-.. automodule:: actions.railway_trafic.IAPF_rail_load_models
+.. automodule:: actions.railway_traffic.IAPF_rail_load_models
     :members:
     :undoc-members:
     :show-inheritance:
@@ -13,7 +14,7 @@ IAPF rail load models
 SIA rail load models
 --------------------
 
-.. automodule:: actions.railway_trafic.SIA_rail_load_models
+.. automodule:: actions.railway_traffic.SIA_rail_load_models
     :members:
     :undoc-members:
     :show-inheritance:
