+++ doc/sphinxDoc/source/preprocessor/python_modules/actions.roadway_traffic.rst
@@ -1,12 +1,11 @@
-Railway traffic
+Roadway traffic
 ===============
-
 
 
 IAPF rail load models
 ---------------------
 
-.. automodule:: actions.railway_trafic.IAPF_rail_load_models
+.. automodule:: actions.railway_traffic.IAPF_rail_load_models
     :members:
     :undoc-members:
     :show-inheritance:
@@ -14,7 +13,7 @@ IAPF rail load models
 SIA rail load models
 --------------------
 
-.. automodule:: actions.railway_trafic.SIA_rail_load_models
+.. automodule:: actions.railway_traffic.SIA_rail_load_models
     :members:
     :undoc-members:
     :show-inheritance:
