+++ python_modules/actions/pedestrian_traffic/readme.md
@@ -1,4 +1,4 @@
-# Trafic loads verification tests
+# Pedestrian traffic loads.
 
 ## References
