+++ python_modules/setup.py
@@ -61,7 +61,7 @@ awc_nds_section_data= ['materials/awc_nds/aux/pr_400_i_joists.json']
 setup(name='XCModules',
       version='0.12.0',
       author='Luis C. Pérez Tato',
-      packages=['geom_utils', 'geotechnics', 'geotechnics.foundations', 'geotechnics.tuneles', 'rough_calculations','solution', 'model', 'model/boundary_cond','model/geometry', 'model/mesh','model/sets','postprocess','import_export','import_export/sciaXML','import_export/sciaXML/xml_basics','import_export/sciaXML/scia_loads','postprocess/gnuplot','postprocess/config','postprocess/reports','postprocess/elements','postprocess/quick_inquiry',"actions","actions/basic_loads","actions/quake","actions/earth_pressure","actions/pedestrian_trafic","actions/railway_trafic","actions/roadway_trafic","actions/snow","actions/weight","actions/wind","actions/thermal","actions/prestress","actions/imposed_strain",'actions/load_combination_utils', 'actions/utils','materials',"materials/aci","materials/aisi","materials/astm_aisc","materials/ec2","materials/ec3","materials/ec5","materials/ceb","materials/ehe","materials/eae","materials/sections","materials/sections/fiber_section","materials/sections/structural_shapes","materials/sections/structural_shapes/aux","materials/sia262","materials/sia263","materials/eota","materials/prestressing","materials/awc_nds", "materials/awc_nds/aux","materials/tm5_809_3","materials/tie_bars", "import_export", "serviceability_limit_states", "postprocess/xcVtk","postprocess/xcVtk/fields","postprocess/xcVtk/diagrams","postprocess/xcVtk/FE_model","postprocess/xcVtk/CAD_model", "connections/steel_connections","misc","misc/latex","misc/sqlite_utils","misc/vtk_utils","misc_utils"],
+      packages=['geom_utils', 'geotechnics', 'geotechnics.foundations', 'geotechnics.tuneles', 'rough_calculations','solution', 'model', 'model/boundary_cond','model/geometry', 'model/mesh','model/sets','postprocess','import_export','import_export/sciaXML','import_export/sciaXML/xml_basics','import_export/sciaXML/scia_loads','postprocess/gnuplot','postprocess/config','postprocess/reports','postprocess/elements','postprocess/quick_inquiry',"actions","actions/basic_loads","actions/quake","actions/earth_pressure","actions/pedestrian_traffic","actions/railway_traffic","actions/roadway_traffic","actions/snow","actions/weight","actions/wind","actions/thermal","actions/prestress","actions/imposed_strain",'actions/load_combination_utils', 'actions/utils','materials',"materials/aci","materials/aisi","materials/astm_aisc","materials/ec2","materials/ec3","materials/ec5","materials/ceb","materials/ehe","materials/eae","materials/sections","materials/sections/fiber_section","materials/sections/structural_shapes","materials/sections/structural_shapes/aux","materials/sia262","materials/sia263","materials/eota","materials/prestressing","materials/awc_nds", "materials/awc_nds/aux","materials/tm5_809_3","materials/tie_bars", "import_export", "serviceability_limit_states", "postprocess/xcVtk","postprocess/xcVtk/fields","postprocess/xcVtk/diagrams","postprocess/xcVtk/FE_model","postprocess/xcVtk/CAD_model", "connections/steel_connections","misc","misc/latex","misc/sqlite_utils","misc/vtk_utils","misc_utils"],
       package_data={'': ['*.pickle']},
       data_files=[(pth_to_libs, ['extensions/xc_base.so','extensions/loadCombinations.so','extensions/geom.so','extensions/xcGnuGts.so','extensions/xc.so']), (pth_to_steel_shape_data, steel_shape_data), (pth_to_awc_nds_section_data, awc_nds_section_data)]
      )
